requirements.txt
Flask
Flask-PyMongo
pymongo
gevent
Pillow
opencv-python
numpy
six
coverage
nose